apps/ai-service/pyproject.toml
[project]
name = "supplygraph-ai-service"
version = "0.1.0"
description = "SupplyGraph AI Service - FastAPI backend with LangGraph workflow orchestration"
authors = [
    {name = "SupplyGraph Team", email = "dev@supplygraph.com"},
]
dependencies = [
    # FastAPI Core
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    
    # LangGraph & AI
    "langgraph>=0.2.74",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    
    # Database (Prisma Python Client)
    "prisma>=0.15.0",
    "asyncpg>=0.30.0",
    
    # Redis/Valkey for queues and caching
    "redis>=5.2.0",
    "celery>=5.4.0",
    
    # HTTP & API
    "httpx>=0.28.0",
    "python-multipart>=0.0.12",
    
    # Document Processing
    "docling>=2.7.0",
    
    # Gmail API
    "google-auth>=2.35.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.150.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=24.4.0",
    
    # Testing
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "W", "F", "I", "B", "C4", "ARG", "SIM"]
ignore = ["E501", "W503", "E203"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true